Display formatting updates

+++ pll.h
@@ -1,4 +1,4 @@
-void PLL_Open(int16_t maxClkDevide, int16_t maxLoopDevide, float referenceFreqMHz);
+void PLL_Open(int16_t maxClkDevide, int16_t maxLoopDevide, int8_t referenceFreqMHz);
 void PLL_Close();
 void PLL_Read();
 void PLL_Write(int freqMHz, int freqDecimal);
